// src/pinecone-generated-ts-fetch/admin/models/APIKey.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pinecone Admin API
 * Provides an API for managing a Pinecone organization and its resources, including projects, API keys, organization users and invites, service accounts, and role bindings. 
 *
 * The version of the OpenAPI document: 2026-07
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * The details of an API key, without the secret.
 * @export
 * @interface APIKey
 */
export interface APIKey {
    /**
     * The unique ID of the API key.
     * @type {string}
     * @memberof APIKey
     */
    id: string;
    /**
     * The name of the API key.
     * @type {string}
     * @memberof APIKey
     */
    name: string;
    /**
     * The ID of the project containing the API key.
     * @type {string}
     * @memberof APIKey
     */
    projectId: string;
    /**
     * The roles assigned to the API key.
     * @type {Array<string>}
     * @memberof APIKey
     */
    roles: Array<string>;
}

/**
 * Check if a given object implements the APIKey interface.
 */
export function instanceOfAPIKey(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "projectId" in value;
    isInstance = isInstance && "roles" in value;

    return isInstance;
}

export function APIKeyFromJSON(json: any): APIKey {
    return APIKeyFromJSONTyped(json, false);
}

export function APIKeyFromJSONTyped(json: any, ignoreDiscriminator: boolean): APIKey {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'projectId': json['project_id'],
        'roles': json['roles'],
    };
}

export function APIKeyToJSON(value?: APIKey | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'project_id': value.projectId,
        'roles': value.roles,
    };
}